rename provider module to web3utils and add test spec
Renamed lib/web3Provider.ts to lib/web3utils.ts per review feedback.
Added test/server/web3utilsSpec.ts to verify singleton behavior.
